requirements.txt
boto3
django
django-cors-headers
django-leaflet
django-storages
django-tailwind-cli
djangorestframework
djangorestframework-gis
gunicorn
numpy
pandas
pdf2image
Pillow
psycopg2-binary
PyMuPDF
PyPDF2
pytesseract
python-decouple

requests
tqdm

pytest
pytest-cov
pytest-sugar
mixer
hypothesis
pytest-django